+++ kadmium-dmx.Data/Types/Fixtures/DMXChannelData.cs
@@ -7,8 +7,8 @@ namespace kadmium_dmx_data.Types.Fixtures
     public class DMXChannelData : IDMXChannelData
     {
         public string Name { get; set; }
-        public byte Min { get; set; } = byte.MinValue;
-        public byte Max { get; set; } = byte.MaxValue;
-        public ushort Address { get; set; }
+        public int Min { get; set; } = byte.MinValue;
+        public int Max { get; set; } = byte.MaxValue;
+        public int Address { get; set; }
     }
 }
